+++ tools/Propulsion.Tool/Program.fs
@@ -339,13 +339,14 @@ module Project =
         let stats = Stats(TimeSpan.FromMinutes 1., TimeSpan.FromMinutes 5., logExternalStats = dumpStoreStats)
         let sink =
             let maxReadAhead, maxConcurrentStreams = 2, 16
-            let handle (stream : FsCodec.StreamName) (span : Propulsion.Sinks.Event[]) ct = task {
+            let handle (stream : FsCodec.StreamName) (span : Propulsion.Sinks.Event[]) = async {
                 match producer with
                 | None -> ()
                 | Some producer ->
                     let json = Propulsion.Codec.NewtonsoftJson.RenderedSpan.ofStreamSpan stream span |> Newtonsoft.Json.JsonConvert.SerializeObject
-                    return! producer.ProduceAsync(FsCodec.StreamName.toString stream, json) |> Async.Ignore |> Async.startImmediateAsTask ct }
-            Propulsion.Streams.Concurrent.Start(Log.Logger, maxReadAhead, maxConcurrentStreams, handle, stats, stats.StatsInterval, Propulsion.Sinks.Event.storedSize, idleDelay = a.IdleDelay)
+                    do! producer.ProduceAsync(FsCodec.StreamName.toString stream, json) |> Async.Ignore
+                return Propulsion.Sinks.AllProcessed, () }
+            Propulsion.Sinks.Factory.StartConcurrent(Log.Logger, maxReadAhead, maxConcurrentStreams, handle, stats, idleDelay = a.IdleDelay)
         let source =
             let nullFilter _ = true
             match storeArgs with